requirements.txt
# API and Web Framework
fastapi==0.109.2
uvicorn==0.22.0
pydantic==2.6.1
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.15
alembic==1.11.1

# Data Analysis
pandas==2.2.1
numpy==1.26.4

# Scheduling
apscheduler==3.10.1

# HTTP Client
httpx==0.24.1

# Visualization
plotly==5.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pyyaml==6.0.1

# Development dependencies
black==23.3.0
flake8==6.0.0
isort==5.12.0
python-dotenv==1.0.0

# Monitoring
prometheus-client==0.19.0
prometheus-fastapi-instrumentator==6.1.0 